+++ cypress/e2e/features/your-voyage-pagination.feature
@@ -9,9 +9,9 @@ Feature: As a shipping agent I can able to report voyage and easily look through
     Then there should be no pagination
     Examples:
       | emailAddress                                       | password      |
-      | abf4b167-04bc-4ff1-b4b0-7f9ceb0b6ffd@mailslurp.com | Test-NMSW-Dev |
+      | 0e25b97b-35bb-4959-8da9-e0439c9a0df9@mailslurp.com | Test-NMSW-Dev |
 
-  Scenario Outline: I can see a page link for each 100 records for account more than 100 voyages
+  Scenario Outline: I can see a page link for each 50 records for account more than 100 voyages
     When I have entered a correct '<emailAddress>' and '<password>' and sign in
     Then I am taken to your-voyages page
     Then I am able to see the pagination based on number of reports
