+++ cypress/support/step_definitions/your-voyage-pagination.steps.js
@@ -2,7 +2,7 @@ import {Then, When} from "@badeball/cypress-cucumber-preprocessor";
 import DeclarationPage from "../../e2e/pages/declaration.page";
 import fileUploadPage from "../../e2e/pages/file-upload.page";
 import SignInPage from "../../e2e/pages/sign-in.page";
-import YourVoyagePage from "../../e2e/pages/your-voyage.page";
+import YourVoyagePage from "../../e2e/pages/your-voyage-page";
 
 Then('I can see the draft details of the voyage, I have uploaded', () => {
   DeclarationPage.checkVoyageDetailsStatus('draft');
